Add files via upload
Parking Management System developed using a stack data structure. The system allows users to perform operations such as selecting the vehicle type (car or motorbike), assigning a unique block number to each vehicle, and calculating revenue based on the vehicle type. It also includes the ability to search for and remove specific vehicles by their license plate, making it an efficient and organized solution for managing parking space. The project showcases the practical application of stacks in solving real-world problems.